agents/phone-calls/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# ElevenLabs SDK
elevenlabs>=2.16.0

# HTTP client for webhooks
httpx>=0.25.0

# Date/time handling
python-dateutil>=2.8.2

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
